fix(editor): prevent issues via sidebar backdrop
There was a couple of things that went wrong when the secondary sidebar
was opened and the user clicked outside it:

i.
When clicking on the console or the boardviewer, the click was not
registered by the sidebar component, because those elements do not
propagate it, so the sidebar would not close.

ii.
When clicking on the monaco editor, the browser would completely freeze.
I was unable to debug that because I could not even profile it to see
what was taking 100% CPU.