feat(editor): make the monaco scrollbars thinner
This is in reference to #92. This patch makes the scrollbars thinner,
but the issue reported, I believe, was mostly about the "overview
ruler", which is a component in the Monaco Editor that's used to
highlight things like the cursor position and the find results.

In the dark theme, the ruler has a left border that is too obtrusive.
Luckily there's a merged PR in vscode that addresses the issue, by
allowing the user (that's us!) to change the color or hide the overview
ruler completely, which is what I would like.

This can be implemented as soon as a version of monaco-editor containing
that patch is released.

References:
https://github.com/Microsoft/vscode/issues/5099
https://github.com/Microsoft/vscode/pull/21459