feat: disconnect other tabs when making a new connection
This allows for only one connection at a time. It would've been great to
support multiple concurrent connections, and it *was* working when you'd
connect different devices, but things would break apart when attempting
multiple connections to the same device. The reason being that there is
no way to differenciate between devices with the same
vendorId/productId.