fix(build): better organize SCSS files and remove stray CSS files
The issue with rebuilding the SCSS during `npm start`, that was fixed by
c013aaf, was actually because of stray CSS files coming from the latest
merge of angular-seed from upstream. That was not a proper fix, but just
a workaround, and I shall revert that patch next.

Additionally, remove unneeded files that belong to upstream angular-seed
but we don't need, and end up messing up our build system.