feat: use a global connection and global terminal
We are effectively only allowing one connection at a time, so there's no
point in leaving 'Connect' buttons tied to editor tabs, nor having
multiple consoles each tied to a tab.

This patch simply centralizes the connection handling in WebUsbService
and the terminal in AppDataService. It adds a 'Connect' button to the
sidebar and as a final nice touch, a glowing green/yellow LED next to
the app title at the top, when there's an active WebUSB connection.