fix(webusb): remove unnecessary check for `device.opened`
We used to need that check when trying to deal with multiple connected
devices at the same time, but now that we've dropped that feature, we
don't need that check anymore. Moreover, that check has the unintended
consequence of preventing a connection if a device was previously opened
from a different browser instance and that instance was closed without
disconnecting.