feat(editor): add STOP button
There is no enabling/disabling the run and stop buttons according to
whether we just begun running or we just stopped. I think we could do
something but it's pointless to try and get too clever only to be always
be missing some cases because there's really no "run status report"
coming from ashell. So let's leave it at the fact that the run and stop
buttons are simple mappings to the ashell console commands and be happy
with it.

Closes #147